genJson: initial commit
genJson is a tool generating the defaultParams.json files for tasks.
It handles auto-detecting the different component of a task, their
dependencies, but also detecting the various dependencies which will be
required by the solutions.

(Note: this version is still quite specific to France-IOI task structure, a
more flexible version is coming.)