remoteGrader: uses the graderqueue for evaluation
The remoteGrader uses a graderqueue instance to perform an evaluation. It can
be an in-place replacement for the taskgrader anywhere: it accepts the job JSON
on its stdin and outputs the result on stdout.
It needs to be configured with the URL to the graderqueue instance API, and a
valid username/password pair for that graderqueue.

'taskstarter remotetest solution.c' uses remoteGrader to perform its test
evaluation of a solution. It uses makeStandaloneJson to generate a complete
JSON, so that the remote graderserver doesn't need to load any files; they're
all included into the JSON.

Check https://github.com/France-ioi/graderqueue for more information.