Example taskSolchecker
Checker has access to the solution source and uses it to grade accordingly.
It counts the number of non-whitespace characters of the solution source, and
if the source is too long, diminishes the grade accordingly.

sol-long-c.c is an example of solution which doesn't get a grade of 100 (but 93
instead) because it is too long.